Correcting metasploit creds csv and adding ipmi
The metasploit creds csv parser was incorrectly assuming a smb hash so
it now checks for the type field for hashes instead.

Only these hash types are 'known' by Kvasir. Metasploit doesn't seem to
have a strict setting for creds' :sname field so we'll have to grow this
as they become more known:

- smb
- rakp_hmac_sha1_hash
- smb_challenge

The JTR POT support also poses a number of challenges as there is no
unified way for most hash types to be stored. JTR seems to use:

$hashtype$hash$salt:password

or

$hashtype$salt$hash:password

And this may not correspond to how MSF or any other tool/user places
hashes inside Kvasir. So we'll kludge along with square tires for now.