Handle sev/cvss with None without erroring and remove IS_EMPTY_OR for f_cvss_score
f_cvss_score and f_severity fields for db.t_vulndata should always be something,
however f_cvss_score was configured with IS_EMPTY_OR() which meant it could be None.
This throws errors because we force casts in some parts of the code and you can't
do that with None.