Nessus6 API Support
This adds support for the Nessus 6 REST API and is based on various
code found on Github. Thanks to everyone who has worked on this in
the past, especially @jfalken

kvasir.yaml settings to use Nessus 6 include:

```
nessus:
  servers: [
    server_name: {
      url: 'https://localhost:8834/',
      secret_key: ''',
      access_key: '',
      username: '',
      password: '',
      version: 6,
      verify_ssl: false
    },
  ]
```

where:

 * secret_key / access_key: your API keys, remove if not used
 * username / password: your username and password, remove if not used
 * verify_ssl: true if you have a trusted certificate, otherwise false

If you use API key, do not set username/password and vice-versa.